+++ examples/host_db18c/cloudinit/post_config_db_env.sh
@@ -6,9 +6,10 @@
 # Name.......: post_config_db_env.sh
 # Author.....: Stefan Oehrli (oes) stefan.oehrli@trivadis.com
 # Editor.....: Stefan Oehrli
-# Date.......: 2021.06.14
+# Date.......: 2021.06.16
 # Revision...: 
-# Purpose....: Script to configure the DB server after bootstrap
+# Purpose....: Script to configure the DB server after bootstrap of an 
+#              initialized system
 # Notes......: --
 # Reference..: --
 # License....: Apache License Version 2.0, January 2004 as shown
@@ -24,7 +25,7 @@ HOST=${HOST:-$(hostname)}
 export SCRIPT_NAME=$(basename $0)               # script name
 export SCRIPT_BIN_DIR=$(dirname $0)             # script bin directory
 # define logfile and logging
-export LOG_BASE=${LOG_BASE:-"/tmp"}                          # Use script directory as default logbase
+export LOG_BASE=${LOG_BASE:-"$SCRIPT_BIN_DIR"}  # Use script directory as default logbase
 # Define Logfile but first reset LOG_BASE if directory does not exists
 if [ ! -d ${LOG_BASE} ] || [ ! -w ${LOG_BASE} ] ; then
     echo "INFO : set LOG_BASE to /tmp"
@@ -55,6 +56,7 @@ else
     echo "WARN: could not source db environment"
 fi
 
-
 echo "INFO: Finish post bootstrap lab environment configuration on host $(hostname) at $(date)"
+echo "INFO: Initiate system reboot in about 1m ----------------------------------"
+sudo shutdown --reboot --no-wall +1
 # --- EOF ---------------------------------------------------------------------